No issue: move relevant :mozilla-detekt-rules to perf/ pkg.
